requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.9.1
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
colorama==0.4.6
distro==1.9.0
Django==5.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.8.0
django-otp==1.6.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.7
langchain-openai==1.1.7
langsmith==0.6.6
openai==2.16.0
orjson==3.11.6
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
zstandard==0.25.0
